+++ client/analytics/report/index.js
@@ -19,7 +19,6 @@ import {
 	getFilteredCurrencyInstance,
 } from '../../lib/currency-context';
 import getReports from './get-reports';
-import './navigation';
 
 /**
  * The Customers Report will not have the `report` param supplied by the router/
